# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Financeiro Backend API",
    "version": "1.0.0",
    "description": "API para gerenciamento financeiro de imóveis, categorias e lançamentos."
  },
  "servers": [
    {
      "url": "https://backend-financeiro-m4r6.onrender.com"
    }
  ],
  "paths": {
    "/imoveis": {
      "get": {
        "summary": "Listar todos os imóveis",
        "operationId": "getImoveis",
        "responses": {
          "200": {
            "description": "Lista de imóveis"
          }
        }
      },
      "post": {
        "summary": "Adicionar novo imóvel",
        "operationId": "addImovel",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": {
                    "type": "string"
                  },
                  "vendido": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "nome",
                  "vendido"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Imóvel criado"
          }
        }
      }
    },
    "/imoveis/{imovel_id}": {
      "get": {
        "summary": "Buscar imóvel por ID",
        "parameters": [
          {
            "name": "imovel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Imóvel encontrado"
          },
          "404": {
            "description": "Imóvel não encontrado"
          }
        }
      },
      "patch": {
        "summary": "Atualizar imóvel",
        "parameters": [
          {
            "name": "imovel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Imóvel atualizado"
          },
          "404": {
            "description": "Imóvel não encontrado"
          }
        }
      },
      "delete": {
        "summary": "Excluir imóvel",
        "parameters": [
          {
            "name": "imovel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Imóvel excluído"
          }
        }
      }
    },
    "/categorias": {
      "get": {
        "summary": "Listar categorias",
        "responses": {
          "200": {
            "description": "Lista de categorias"
          }
        }
      },
      "post": {
        "summary": "Adicionar nova categoria",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "categoria": {
                    "type": "string"
                  },
                  "dc": {
                    "type": "string"
                  }
                },
                "required": [
                  "categoria",
                  "dc"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Categoria criada"
          }
        }
      }
    },
    "/categorias/{categoria_id}": {
      "delete": {
        "summary": "Excluir categoria",
        "parameters": [
          {
            "name": "categoria_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Categoria excluída"
          }
        }
      }
    },
    "/dashboard/lancamentos/incompletos/{id_imovel}": {
      "get": {
        "summary": "Listar lançamentos incompletos de um imóvel",
        "parameters": [
          {
            "name": "id_imovel",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de lançamentos incompletos"
          }
        }
      }
    },
    "/dashboard/lancamentos/completos/{id_imovel}": {
      "get": {
        "summary": "Listar lançamentos completos de um imóvel",
        "parameters": [
          {
            "name": "id_imovel",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de lançamentos completos"
          }
        },
        "operationId": "getLancamentosCompletosPorImovel"
      }
    },
    "/dashboard/lancamentos/lote": {
      "post": {
        "summary": "Adicionar lançamentos em lote",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "string"
                    },
                    "descricao": {
                      "type": "string"
                    },
                    "valor": {
                      "type": "number"
                    },
                    "id_imovel": {
                      "type": "integer"
                    },
                    "id_categoria": {
                      "type": "integer"
                    },
                    "id_situacao": {
                      "type": "integer"
                    },
                    "ativo": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "data",
                    "descricao",
                    "valor",
                    "id_imovel"
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Lançamentos adicionados"
          }
        }
      }
    },
    "/analise/lancamentos": {
      "get": {
        "summary": "Obter todos os lançamentos detalhados para análise",
        "operationId": "getLancamentosUnificados",
        "responses": {
          "200": {
            "description": "Lista de lançamentos completos com dados relacionais"
          }
        }
      }
    }
  }
}